+++ internal/repository/document_version_interface.go
@@ -12,6 +12,6 @@ type DocumentVersionRepository interface {
 	ListByDocument(ctx context.Context, userID, documentID string) ([]entity.DocumentVersion, error)
 	FindByID(ctx context.Context, userID, documentID, versionID string) (entity.DocumentVersion, bool, error)
 	FindLatestByDocument(ctx context.Context, userID, documentID string) (entity.DocumentVersion, bool, error)
-	SaveVersionAndReindex(ctx context.Context, doc entity.Document, job *entity.DocumentProcessingJob, version *entity.DocumentVersion, chunks []entity.DocumentChunk, readyStatus, successJobStatus int16, finishedAt time.Time) error
-	ReindexVersion(ctx context.Context, doc entity.Document, job *entity.DocumentProcessingJob, version entity.DocumentVersion, chunks []entity.DocumentChunk, readyStatus, successJobStatus int16, finishedAt time.Time) error
+	SaveVersionAndReindex(ctx context.Context, doc entity.Document, job *entity.DocumentProcessingJob, version *entity.DocumentVersion, chunks []entity.DocumentChunk, readyStatus, successJobStatus int, finishedAt time.Time) error
+	ReindexVersion(ctx context.Context, doc entity.Document, job *entity.DocumentProcessingJob, version entity.DocumentVersion, chunks []entity.DocumentChunk, readyStatus, successJobStatus int, finishedAt time.Time) error
 }
